requirements.txt
numpy
scipy
torch
torchvision
opencv-python
pillow
tqdm
ultralytics
pytesseract
